:bug: Fix HTML content not rendering in the compiled version
Phew, that was a tough one! The HTML content was rendering fine when
running Hook Manager in Eclipse's debug mode,  but not when compiling it
into a Jar.

Now it will work as long as the HTML folder is placed next to the Jar
instead of being inside the Jar.